multiple login prevention for student only and MarkingCSV export added

+++ WebApp/models.py
@@ -214,17 +214,20 @@ class UserSession(models.Model):
 
 @receiver(user_logged_in)
 def remove_other_sessions(sender, user, request, **kwargs):
-    # remove other sessions
-    Session.objects.filter(usersession__user=user).delete()
+    if user.Is_Teacher or user.Is_CenterAdmin:
+        pass
+    else:
+        # remove other sessions
+        Session.objects.filter(usersession__user=user).delete()
 
-    # save current session
-    request.session.save()
+        # save current session
+        request.session.save()
 
-    # create a link from the user to the current session (for later removal)
-    UserSession.objects.get_or_create(
-        user=user,
-        session=Session.objects.get(pk=request.session.session_key)
-    )
+        # create a link from the user to the current session (for later removal)
+        UserSession.objects.get_or_create(
+            user=user,
+            session=Session.objects.get(pk=request.session.session_key)
+        )
 
 class CourseInfo(models.Model):
     Course_Name = CharField(max_length=240)
